move some deps to devDeps. remove legacy context
