+++ shared/common-adapters/relative-popup-hoc.desktop.js
@@ -4,7 +4,7 @@ import * as React from 'react'
 import {includes, throttle, without} from 'lodash-es'
 import Box from './box'
 import ReactDOM, {findDOMNode} from 'react-dom'
-import EscapeHandler from '../util/escape-handler'
+import EscapeHandler from '../util/escape-handler.desktop'
 import {connect} from '../util/container'
 import {type StylesCrossPlatform, collapseStyles} from '../styles'
 import type {Position, RelativePopupHocType, Props} from './relative-popup-hoc.types'
