+++ shared/package.json
@@ -127,9 +127,6 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
-    "@babel/plugin-proposal-nullish-coalescing-operator": "7.0.0",
-    "@babel/plugin-proposal-optional-catch-binding": "7.0.0",
-    "@babel/plugin-proposal-optional-chaining": "7.0.0",
     "applescript": "1.0.0",
     "base64-js": "1.3.0",
     "buffer": "5.2.1",
@@ -140,7 +137,6 @@
     "emoji-datasource-apple": "4.1.0",
     "emoji-mart": "2.8.1",
     "exif-js": "2.3.0",
-    "flow-watch": "1.1.3",
     "framed-msgpack-rpc": "1.1.15",
     "getenv": "0.7.0",
     "glamor": "2.20.40",
@@ -157,7 +153,6 @@
     "mousetrap": "1.6.2",
     "path-parse": "1.0.6",
     "prettier": "1.14.3",
-    "prop-types": "15.6.2",
     "purepack": "1.0.4",
     "qrcode-generator": "git://github.com/keybase/qrcode-generator#keybase-changes-off-140",
     "re-reselect": "2.1.0",
@@ -187,7 +182,6 @@
     "rn-fetch-blob": "0.10.13",
     "semver": "5.5.1",
     "shallowequal": "1.1.0",
-    "terser-webpack-plugin": "1.1.0",
     "tlds": "1.203.1",
     "typedarray-to-buffer": "3.1.5",
     "url-parse": "1.4.3",
@@ -195,6 +189,10 @@
     "uuid": "3.3.2"
   },
   "devDependencies": {
+    "terser-webpack-plugin": "1.1.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "7.0.0",
+    "@babel/plugin-proposal-optional-catch-binding": "7.0.0",
+    "@babel/plugin-proposal-optional-chaining": "7.0.0",
     "@babel/cli": "7.1.2",
     "@babel/core": "7.1.2",
     "@babel/node": "7.0.0",
@@ -245,11 +243,9 @@
     "net": "git://github.com/keybase/nullModule",
     "null-loader": "0.1.1",
     "pegjs": "0.10.0",
-    "raf": "3.4.0",
     "react-hot-loader": "4.3.11",
     "react-test-renderer": "16.6.0-alpha.8af6728",
     "redux-logger": "3.0.6",
-    "source-map-support": "0.5.9",
     "style-loader": "0.23.0",
     "tls": "git://github.com/keybase/nullModule",
     "webfonts-generator": "0.4.0",
@@ -260,7 +256,6 @@
     "webpack-merge": "4.1.4"
   },
   "optionalDependencies": {
-    "flow-coverage-report": "0.6.0",
     "regedit": "2.2.7"
   },
   "resolutions": {
