--- shared/util/escape-handler.js.flow
@@ -1,15 +1,0 @@
-// @flow
-import * as React from 'react'
-
-export type GlobalProps = {
-  children?: React.Node,
-}
-
-export class GlobalEscapeHandler extends React.Component<GlobalProps> {}
-
-export type Props = {
-  onESC: ?() => void,
-  children?: React.Node,
-}
-
-export default class EscapeHandler extends React.Component<Props> {}
